Refactor explainer classes and update dataset handling
- Renamed `Explainer` class to `ExplainerABC` across the codebase for clarity and consistency.
- Updated imports to reflect the new class name in all relevant files.
- Modified the `DataInfo` class to ensure proper handling of tensor types by converting to float where necessary.
- Added support for the "HIV" dataset in the dataset utility functions, including appropriate train-test splitting and feature handling.
- Introduced a new `GNNExplainerWrap` class in both node and graph level explainers to encapsulate GNN explanation logic.
- Adjusted the forward methods in the GCN model to ensure tensor types are correctly managed.
- Enhanced the training loop to handle tensor shapes and types appropriately, ensuring compatibility with the updated model outputs.

+++ src/datasets/dataset.py
@@ -8,7 +8,7 @@ class DataInfo:
     def __init__(self, cfg: OmegaConf, data: Dataset) -> None:
         self.device = "cuda" if torch.cuda.is_available() and cfg.device == "cuda" else "cpu"
         self.data = data
-        self.distribution_mean = torch.mean(self.data.x, dim=0).cpu() if cfg.task.name == "Node" else torch.mean(self.data.dataset.data.x, dim=0).cpu()
+        self.distribution_mean = torch.mean(self.data.x, dim=0).cpu() if cfg.task.name == "Node" else torch.mean(self.data.dataset.data.x.float(), dim=0).cpu()
         self._distribution_mean_projection = None
         self.kfold = None
         self.inv_covariance_matrix = None
